renaming stream.py to module.py
