+++ nexus_constructor/add_component_window.py
@@ -26,7 +26,7 @@ from nexus_constructor.model.geometry import (
     OFFGeometryNoNexus,
 )
 from nexus_constructor.model.model import Model
-from nexus_constructor.model.stream import Link
+from nexus_constructor.model.module import Link
 from nexus_constructor.pixel_options import PixelOptions
 from nexus_constructor.ui_utils import (
     file_dialog,
