+++ nexus_constructor/model/helpers.py
@@ -3,7 +3,7 @@ from typing import TYPE_CHECKING, Any, List, Union
 if TYPE_CHECKING:
     from nexus_constructor.model.attributes import Attributes
     from nexus_constructor.model.group import Group
-    from nexus_constructor.model.stream import FileWriterModule
+    from nexus_constructor.model.module import FileWriterModule
 
 
 def __find_item_index(list_to_look_in: List[Any], item_name: str):
