+++ tests/model/test_component_fields.py
@@ -1,7 +1,7 @@
 import numpy as np
 
 from nexus_constructor.model.component import Component, add_fields_to_component
-from nexus_constructor.model.stream import Dataset
+from nexus_constructor.model.module import Dataset
 
 
 class DummyListWidget:
