+++ tests/test_component_tree_model.py
@@ -12,7 +12,7 @@ from nexus_constructor.component_tree_model import (
 from nexus_constructor.model.component import Component
 from nexus_constructor.model.entry import Instrument
 from nexus_constructor.model.model import Model
-from nexus_constructor.model.stream import Dataset
+from nexus_constructor.model.module import Dataset
 from nexus_constructor.model.value_type import ValueTypes
