+++ tests/test_link_transformation.py
@@ -1,7 +1,7 @@
 from PySide2.QtGui import QVector3D
 
 from nexus_constructor.model.component import Component
-from nexus_constructor.model.stream import Dataset
+from nexus_constructor.model.module import Dataset
 from nexus_constructor.model.value_type import ValueTypes
